Fixed bug when joining answers + theme update
1. Theme has been updated to bootstrap version 4 using the `bslib` package.
2. Google font properly imported using the `bslib` package.
3. Changed joining function from `full_join()` to `merge()`. The latter doesn't
enforce strict type checking, which works well when there is type mismatch
between the old answers and the new user input. At the worst case scenario,
every variable would be coerced to a character vector, which in context doesn't
matter because it's all uploaded to google sheets anyway.
4. Fixed problem where unanswered multiple choice questions would return `NULL`,
which caused errors when trying to coerce user answers list with a `NULL`
element to a `data frame`